% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{default_validate_dag}
\alias{default_validate_dag}
\title{Validate a directed acyclic graph for use with rollup}
\usage{
default_validate_dag(dag)
}
\arguments{
\item{dag}{An igraph directed acyclic graph}
}
\value{
TRUE if valid, stops otherwise
}
\description{
Validate a directed acyclic graph for use with rollup
}
\examples{
default_validate_dag(test_dag)
}
