% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{default_validate_tree}
\alias{default_validate_tree}
\title{Validate a tree for use with \code{rollup()}}
\usage{
default_validate_tree(tree)
}
\arguments{
\item{tree}{\code{igraph} directed graph that is a valid single-rooted in-tree
and whose vertex names are keys from the data set}
}
\value{
single root vertex identifier if tree is valid; stops otherwise
}
\description{
\code{default_validate_tree()} ensures that a tree is acyclic,
loop-free, single-edged, connected, directed, and single-rooted with edge
direction from child to parent.
}
\examples{
default_validate_tree(wbs_tree)
}
