% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{df_get_by_id}
\alias{df_get_by_id}
\title{Get property by key "id" from data frame}
\usage{
df_get_by_id(df, idval, prop)
}
\arguments{
\item{df}{A data frame}

\item{idval}{ID of the row to get}

\item{prop}{Name of the column to get}
}
\value{
The requested value
}
\description{
\code{df_get_by_id} returns the value of specified property (column) in a specified row
of a data frame. The row is specified by a value for the \code{id} column.
}
\examples{
df_get_by_id(wbs_table, "1.1", "work")
}
