% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{df_get_by_key}
\alias{df_get_by_key}
\title{Get property by key from data frame}
\usage{
df_get_by_key(df, key, keyval, prop)
}
\arguments{
\item{df}{A data frame}

\item{key}{Name of the column used as key}

\item{keyval}{Value of the key for the specified row}

\item{prop}{Column name of the property value to get}
}
\description{
\code{df_get_by_key} returns the value of specified property (column) in a specified row
of a data frame. The row is specified by a key column and a value from that column.
}
\examples{
df_get_by_key(wbs_table, "id", "1", "work")
}
