% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{df_get_keys}
\alias{df_get_keys}
\title{Get keys from a data frame}
\usage{
df_get_keys(df, key)
}
\arguments{
\item{df}{A data frame}

\item{key}{Name of the column used as key}
}
\value{
All values of the key column
}
\description{
\code{df_get_keys} gets all values from a designated column in a data frame.
}
\examples{
df_get_keys(wbs_table, "id")
}
