% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{df_get_row_by_key}
\alias{df_get_row_by_key}
\title{Get row by key from data frame}
\usage{
df_get_row_by_key(df, key, keyval)
}
\arguments{
\item{df}{a data frame}

\item{key}{name of the column used as key}

\item{keyval}{value of the key for the specified row}
}
\value{
A named list of values from the requested row
}
\description{
Get row by key from data frame
}
\examples{
df_get_row_by_key(wbs_table, "id", "1.1")
}
