% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{df_set_by_id}
\alias{df_set_by_id}
\title{Title}
\usage{
df_set_by_id(df, id, prop, value)
}
\arguments{
\item{df}{A data frame}

\item{id}{ID of the row to set}

\item{prop}{Name of the column to set}

\item{value}{Value to set}
}
\value{
The updated dataframe
}
\description{
Title
}
