% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{df_set_by_key}
\alias{df_set_by_key}
\title{Set property by key in data frame}
\usage{
df_set_by_key(df, key, keyval, prop, val)
}
\arguments{
\item{df}{a data frame}

\item{key}{name of the column used as key}

\item{keyval}{value of the key for the specified row}

\item{prop}{column name of the property value to get}

\item{val}{value to set}
}
\value{
The updated data frame
}
\description{
Set property by key in data frame
}
\examples{
df_set_by_key(wbs_table, "id", "1", "work", 45.6)
}
