% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{df_set_row_by_id}
\alias{df_set_row_by_id}
\title{Set row by key "id"  in data frame}
\usage{
df_set_row_by_id(df, idval, list)
}
\arguments{
\item{df}{a data frame}

\item{idval}{id of the specified row}

\item{list}{named list of values to set}
}
\value{
The updated data frame
}
\description{
Set row by key "id"  in data frame
}
\examples{
l <- list(id = "1.1", pid = "1", name = "Thermal", work = 11.9, budget = 25001)
df_set_row_by_id(wbs_table, "1.1", l)
}
