% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{df_set_row_by_key}
\alias{df_set_row_by_key}
\title{Set row by key in data frame}
\usage{
df_set_row_by_key(df, key, keyval, list)
}
\arguments{
\item{df}{a data frame}

\item{key}{name of the column used as key}

\item{keyval}{value of the key for the specified row}

\item{list}{named list of values to set}
}
\value{
The updated data frame
}
\description{
Set row by key in data frame
}
\examples{
l <- list(id = "1.1", pid = "1", name = "Thermal", work = 11.9, budget = 25001)
df_set_row_by_key(wbs_table, "id", "1.1", l)
}
