% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fault_table.R
\docType{data}
\name{fault_table}
\alias{fault_table}
\title{Example Fault Tree Data}
\format{
A data frame with columns:
\describe{
\item{id}{unique key for each row}
\item{type}{event type ("basic", "and", or "or")}
\item{prob}{event probability}
}
}
\source{
\url{https://control.com/technical-articles/deep-dive-into-fault-tree-analysis/}
}
\usage{
fault_table
}
\description{
Example Fault Tree Data
}
\keyword{datasets}
