% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fault_tree.R
\docType{data}
\name{fault_tree}
\alias{fault_tree}
\title{Example Fault Tree}
\format{
An igraph tree with edges from child id to parent id.
}
\source{
\url{https://control.com/technical-articles/deep-dive-into-fault-tree-analysis/}
}
\usage{
fault_tree
}
\description{
Example Fault Tree
}
\keyword{datasets}
