% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dag.R
\docType{data}
\name{test_dag}
\alias{test_dag}
\title{Example Directed Acyclic Graph}
\format{
An igraph DAG with edges from child id to parent id.
}
\usage{
test_dag
}
\description{
Example Directed Acyclic Graph
}
\keyword{datasets}
