% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{update_df_prop_by_key}
\alias{update_df_prop_by_key}
\title{Update a property in a dataframe}
\usage{
update_df_prop_by_key(df, key, target, sources, prop, ...)
}
\arguments{
\item{df}{a data frame}

\item{key}{name of the column serving as key}

\item{target}{key of data set element to be updated}

\item{sources}{keys of data set elements to be combined}

\item{prop}{column name of the property}

\item{...}{other arguments passed to \code{update_prop()}}
}
\value{
The updated data frame
}
\description{
\code{update_df_prop_by_key()} is a convenience wrapper around \code{update_prop()}
for the common case in which the data set is a data frame.
}
\examples{
update_df_prop_by_key(wbs_table, "id", "1", list("1.1", "1.2"), "work")
}
