% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{update_prop}
\alias{update_prop}
\title{Update a Data Set With Recursively-Defined Properties}
\usage{
update_prop(
  ds,
  target,
  sources,
  set,
  get,
  combine = function(l) Reduce("+", l),
  override = function(ds, target, v) v
)
}
\arguments{
\item{ds}{Data set to be updated}

\item{target}{Key of data set element to be updated}

\item{sources}{Keys of data set elements to be combined}

\item{set}{Method to set properties for a target element}

\item{get}{Method to get properties for source elements}

\item{combine}{Method to combine properties}

\item{override}{Method to selectively override combine() results}
}
\value{
Updated data set
}
\description{
\code{update_prop()} calls user-specified methods to get properties
of a source set of elements in a data set, combine those properties,
and set the properties of a target element to the combined value.
If the source set is empty, the data set is returned unmodified. The
default combine operation is addition.
}
\details{
The \code{override} argument can be used to selectively override the
computed value based on the target element. By default, it simply
returns the value computed by combine().
}
\examples{
update_prop(wbs_table, "1", list("1.1", "1.2"),
  function(d, k, v) {d[d$id == k, "work"] <- v; d},
  function(d, k) d[d$id == k, "work"]
 )
update_prop(wbs_table, "1", list("1.1", "1.2"),
  function(d, k, v) {d[d$id == k, c("work", "budget")] <- v; d},
  function(d, k) d[d$id == k, c("work", "budget")],
  function(l) Reduce("+", l)
)
}
