% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{validate_df_by_id}
\alias{validate_df_by_id}
\title{Validate a dataframe with key "id" for rollup()}
\usage{
validate_df_by_id(tree, df, prop, ...)
}
\arguments{
\item{tree}{The tree to validate against}

\item{df}{A datafame}

\item{prop}{Property whose value is checked (leaf elements only)}

\item{...}{Other parameters passed to validate_ds()}
}
\value{
TRUE if validation succeeds, halts otherwise
}
\description{
\code{validate_df_by_id()} is a convenience wrapper for \code{validate_ds()} for the common case in which the
data set is a dataframe with key column named "id".
}
\examples{
validate_df_by_id(wbs_tree, wbs_table, "work")
}
