% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{validate_df_by_key}
\alias{validate_df_by_key}
\title{Validate a data frame For rollup()}
\usage{
validate_df_by_key(tree, df, key, prop, ...)
}
\arguments{
\item{tree}{tree to validate against}

\item{df}{data frame}

\item{key}{name of the column serving as key}

\item{prop}{property whose value is checked (leaf elements only)}

\item{...}{other parameters passed to validate_ds()}
}
\value{
TRUE if validation succeeds, halts otherwise
}
\description{
\code{validate_df_by_key()} is a convenience wrapper for \code{validate_ds()} for the common case in which the
data set is a dataframe.
}
\examples{
validate_df_by_key(wbs_tree, wbs_table, "id", "work")
}
