% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbs_table.R
\docType{data}
\name{wbs_table}
\alias{wbs_table}
\title{Example Work Breakdown Structure Data}
\format{
A data frame with columns:
\describe{
\item{id}{unique key for each row}
\item{pid}{parent key for each row}
\item{name}{character name of the element}
\item{work}{percent of total work for this element}
\item{budget}{budget for this element}
}
}
\source{
\url{https://www.workbreakdownstructure.com}
}
\usage{
wbs_table
}
\description{
Example Work Breakdown Structure Data
}
\keyword{datasets}
