% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbs_tree.R
\docType{data}
\name{wbs_tree}
\alias{wbs_tree}
\title{Example Work Breakdown Structure Data}
\format{
An igraph tree with edges from child id to parent id.
}
\source{
\url{https://www.workbreakdownstructure.com}
}
\usage{
wbs_tree
}
\description{
Example Work Breakdown Structure Data
}
\keyword{datasets}
