% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{default_validate_tree}
\alias{default_validate_tree}
\title{Validate a tree for use with rollup()}
\usage{
default_validate_tree(tree)
}
\arguments{
\item{tree}{An igraph directed graph to be validated}
}
\value{
single root element if tree is valid; stops otherwise
}
\description{
default_validate_tree() ensures that a tree is acyclic,
loop-free, single-edged, connected, directed, and single-rooted with edge
direction from child to parent.
}
