% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{df_get_by_id}
\alias{df_get_by_id}
\title{Title}
\usage{
df_get_by_id(df, id, prop)
}
\arguments{
\item{df}{A data frame}

\item{id}{ID of the row to get}

\item{prop}{Name of the column to get}
}
\value{
The desired value
}
\description{
Title
}
