% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{df_get_by_key}
\alias{df_get_by_key}
\title{Get row by key "id" from data frame}
\usage{
df_get_by_key(df, key, keyval, prop)
}
\arguments{
\item{df}{a data frame}

\item{key}{name of the column used as key}

\item{keyval}{value of the key for the specified row}

\item{prop}{column name of the property value to get}
}
\value{
The requested value
}
\description{
\code{df_get_by_key} returns the value of specified property (column) in a specified row
of a data frame. The row is specified by a key column and a value from that column.
}
\examples{
df_get_by_key(wbs_table, "id", "1.1", "work")
}
