% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{df_get_row_by_id}
\alias{df_get_row_by_id}
\title{Get row by key from data frame}
\usage{
df_get_row_by_id(df, idval)
}
\arguments{
\item{df}{a data frame}

\item{idval}{id of the row to get}
}
\value{
A named list of values from the requested row
}
\description{
Get row by key from data frame
}
\examples{
df_get_row_by_id(wbs_table, "1.1")
}
