% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{df_set_by_id}
\alias{df_set_by_id}
\title{Set property by key "id" in data frame}
\usage{
df_set_by_id(df, idval, prop, val)
}
\arguments{
\item{df}{a data frame}

\item{idval}{id value for the specified row}

\item{prop}{column name of the property value to get}

\item{val}{value to set}
}
\value{
updated data frame
}
\description{
Set property by key "id" in data frame
}
\examples{
df_set_by_id(wbs_table, "1", "work", 45.6)
}
