% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_prop.R
\name{update_df_prop_by_id}
\alias{update_df_prop_by_id}
\title{Update a property in a dataframe with key "id"}
\usage{
update_df_prop_by_id(df, target, sources, prop, ...)
}
\arguments{
\item{df}{A data frame}

\item{target}{Key of data set element to be updated}

\item{sources}{Keys of data set elements to be combined}

\item{prop}{Column name of the property}

\item{...}{Other arguments passed to \code{update_prop()}}
}
\value{
The updated dataframe
}
\description{
\code{update_df_prop_by_id()} is a convenience wrapper around \code{update_prop()}
for the common case in which the data set is a dataframe whose key column
is named "id"
}
\examples{
update_df_prop_by_id(wbs_table, "1", list("1.1", "1.2"), "work")
}
