% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollup.R
\name{validate_ds}
\alias{validate_ds}
\title{Validates a data set for use with \code{rollup()}}
\usage{
validate_ds(tree, ds, get_keys, get_prop, op = function(x) is.numeric(x))
}
\arguments{
\item{tree}{The tree to check against}

\item{ds}{The data set to validate}

\item{get_keys}{Method get keys of the data set called as get_keys(ds)}

\item{get_prop}{Method to get the property to validate for leaf element with id l, called as get_prop(ds, l)}

\item{op}{Logical method to test return value of get_prop() (default \code{is.numeric()}); returns TRUE if OK}
}
\value{
TRUE if validation succeeds, halts otherwise
}
\description{
\code{validate_ds()} ensures that a data set contains the same identifiers as a specified tree and that
elements of the data set corresponding to leaf verticies in the tree satisfy a user-specified predicate.
}
\examples{
validate_ds(wbs_tree, wbs_table, function(d) d$id, function(d, l) d[d$id == l, "work"])

}
